Split provider relayer and local relayer
