+++ packages/relayer/src/index.ts
@@ -31,6 +31,7 @@ export interface Relayer {
 
 export { LocalRelayer } from './local-relayer'
 export { RpcRelayer } from './rpc-relayer'
+export { ProviderRelayer } from './provider-relayer'
 
 export function isRelayer(cand: any): cand is Relayer {
   return cand && cand.estimateGasLimits !== undefined && cand.gasRefundOptions !== undefined &&
