+++ packages/relayer/src/local-relayer.ts
@@ -1,23 +1,16 @@
-import { TransactionResponse, BlockTag } from '@ethersproject/providers'
+import { TransactionResponse } from '@ethersproject/providers'
 import { Signer as AbstractSigner, ethers } from 'ethers'
-import { walletContracts } from '@0xsequence/abi'
-import { SignedTransactions, Transaction, TransactionEncoded } from '@0xsequence/transactions'
+import { SignedTransactions, Transaction } from '@0xsequence/transactions'
 import { WalletContext } from '@0xsequence/network'
-import { WalletConfig, addressOf } from '@0xsequence/config'
-import { BaseRelayer } from './base-relayer'
+import { WalletConfig } from '@0xsequence/config'
 import { Relayer } from '.'
+import { ProviderRelayer } from './provider-relayer'
 
-const DEFAULT_GAS_LIMIT = ethers.BigNumber.from(800000)
+export class LocalRelayer extends ProviderRelayer implements Relayer {
+  constructor(private signer: AbstractSigner) {
+    super(signer.provider!)
 
-export class LocalRelayer extends BaseRelayer implements Relayer {
-  private readonly signer: AbstractSigner
-
-  constructor(signer: AbstractSigner) {
-    super(true, signer.provider)
-    // if (!signer.provider) {
-    //   throw new Error('signer.provider is not set, but is required')
-    // }
-    this.signer = signer
+    if (!signer.provider) throw new Error("Signer must have a provider")
   }
 
   async deployWallet(config: WalletConfig, context: WalletContext): Promise<TransactionResponse> {
@@ -28,50 +21,6 @@ export class LocalRelayer extends BaseRelayer implements Relayer {
 
     // NOTE: for hardhat to pass, we have to set the gasLimit directly, as its unable to estimate
     return this.signer.sendTransaction({ ...walletDeployTxn, gasLimit: ethers.constants.Two.pow(17) } )
-    // return this.signer.sendTransaction(walletDeployTxn)
-  }
-
-  async estimateGasLimits(
-    config: WalletConfig,
-    context: WalletContext,
-    ...transactions: Transaction[]
-  ): Promise<Transaction[]> {
-    const walletAddr = addressOf(config, context)
-
-    const gasCosts = await Promise.all(transactions.map(async tx => {
-      // Respect gasLimit request of the transaction (as long as its not 0)
-      if (tx.gasLimit && !ethers.BigNumber.from(tx.gasLimit || 0).eq(ethers.constants.Zero)) {
-        return tx.gasLimit
-      }
-
-      // Fee can't be estimated locally for delegateCalls
-      if (tx.delegateCall) {
-        return DEFAULT_GAS_LIMIT
-      }
-
-      // Fee can't be estimated for self-called if wallet hasn't been deployed
-      if (tx.to === walletAddr && !(await this.isWalletDeployed(walletAddr))) {
-        return DEFAULT_GAS_LIMIT
-      }
-
-      if (!this.signer.provider) {
-        throw new Error('signer.provider is not set, but is required')
-      }
-
-      // TODO: If the wallet address has been deployed, gas limits can be
-      // estimated with more accurately by using self-calls with the batch transactions one by one
-      return this.signer.provider.estimateGas({
-        from: walletAddr,
-        to: tx.to,
-        data: tx.data,
-        value: tx.value
-      })
-    }))
-
-    return transactions.map((t, i) => {
-      t.gasLimit = gasCosts[i]
-      return t
-    })
   }
 
   async gasRefundOptions(
@@ -80,26 +29,6 @@ export class LocalRelayer extends BaseRelayer implements Relayer {
     ..._transactions: Transaction[]
   ): Promise<Transaction[][]> {
     return [[]]
-  }
-
-  async getNonce(
-    config: WalletConfig,
-    context: WalletContext,
-    space?: number,
-    blockTag?: BlockTag
-  ): Promise<number> {
-    if (!this.provider) {
-      throw new Error('provider is not set')
-    }
-
-    const addr = addressOf(config, context)
-    
-    if ((await this.provider.getCode(addr)) === '0x') {
-      return 0
-    }
-
-    const module = new ethers.Contract(addr, walletContracts.mainModule.abi, this.signer.provider)
-    return (await module.nonce({ blockTag: blockTag })).toNumber()
   }
 
   async relay(signedTxs: SignedTransactions): Promise<TransactionResponse> {
